Module google_deploymentmanager2_beta2::api

Structs

Deployment
Activities

DeploymentDeleteCall
Deletes a deployment and all of the resources in the deployment.

DeploymentGetCall
Gets information about a specific deployment.

DeploymentInsertCall
Creates a deployment and all of the resources described by the deployment manifest.

DeploymentListCall
Lists all deployments for a given project.

DeploymentManager
Central instance to access all DeploymentManager related resource activities

DeploymentMethods
A builder providing access to all methods supported on deployment resources. It is not used directly, but through the DeploymentManager hub.

DeploymentPatchCall
Updates a deployment and all of the resources described by the deployment manifest. This method supports patch semantics.

DeploymentUpdate
This type is not used in any activity, and only used as part of another schema.

DeploymentUpdateCall
Updates a deployment and all of the resources described by the deployment manifest.

DeploymentsListResponse
A response containing a partial list of deployments and a page token used to build the next request if the request has been truncated.

ImportFile
This type is not used in any activity, and only used as part of another schema.

Manifest
Activities

ManifestGetCall
Gets information about a specific manifest.

ManifestListCall
Lists all manifests for a given deployment.

ManifestMethods
A builder providing access to all methods supported on manifest resources. It is not used directly, but through the DeploymentManager hub.

ManifestsListResponse
A response containing a partial list of manifests and a page token used to build the next request if the request has been truncated.

Operation
An Operation resource, used to manage asynchronous API requests.

OperationError
[Output Only] If errors are generated during processing of the operation, this field will be populated.

OperationErrorErrors
[Output Only] The array of errors encountered while processing this operation.

OperationGetCall
Gets information about a specific operation.

OperationListCall
Lists all operations for a project.

OperationMethods
A builder providing access to all methods supported on operation resources. It is not used directly, but through the DeploymentManager hub.

OperationWarnings
[Output Only] If warning messages are generated during processing of the operation, this field will be populated.

OperationWarningsData
[Output Only] Metadata about this warning in key: value format. For example: "data": [ { "key": "scope", "value": "zones/us-east1-d" }

OperationsListResponse
A response containing a partial list of operations and a page token used to build the next request if the request has been truncated.

Resource
Activities

ResourceGetCall
Gets information about a single resource.

ResourceListCall
Lists all resources in a given deployment.

ResourceMethods
A builder providing access to all methods supported on resource resources. It is not used directly, but through the DeploymentManager hub.

ResourceUpdate
This type is not used in any activity, and only used as part of another schema.

ResourcesListResponse
A response containing a partial list of resources and a page token used to build the next request if the request has been truncated.

TargetConfiguration
This type is not used in any activity, and only used as part of another schema.

Type
A resource type supported by Deployment Manager.

TypeListCall
Lists all resource types for Deployment Manager.

TypeMethods
A builder providing access to all methods supported on type resources. It is not used directly, but through the DeploymentManager hub.

TypesListResponse
A response that returns all Types supported by Deployment Manager

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.