Struct google_deploymentmanager2_beta2::Deployment

pub struct Deployment {
    pub update_time: Option<String>,
    pub description: Option<String>,
    pub name: Option<String>,
    pub update: Option<DeploymentUpdate>,
    pub manifest: Option<String>,
    pub state: Option<String>,
    pub intent: Option<String>,
    pub insert_time: Option<String>,
    pub fingerprint: Option<String>,
    pub id: Option<String>,
    pub target: Option<TargetConfiguration>,
}

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

update deployments (request)

get deployments (response)

insert deployments (request)

list deployments (none)

patch deployments (request)

delete deployments (none)

Fields

update_time

[Output Only] Timestamp when the deployment was updated, in RFC3339 text format .

description

An optional user-provided description of the deployment.

name

Name of the resource; provided by the client when the resource is created. The name must be 1-63 characters long, and comply with RFC1035. Specifically, the name must be 1-63 characters long and match the regular expression a-z? which means the first character must be a lowercase letter, and all following characters must be a dash, lowercase letter, or digit, except the last character, which cannot be a dash.

update

[Output Only] If Deployment Manager is currently updating or previewing an update to this deployment, the updated configuration appears here.

manifest

[Output Only] URL of the manifest representing the last manifest that was successfully deployed.

state

[Output Only] The current state of the deployment. This can be DEPLOYED, DEPLOYMENT_FAILED, PREVIEWING, UPDATING, and CANCELING.

intent

[Input Only] Specifies how Deployment Manager should apply this template. Possible options are PREVIEW, UPDATE, and CANCEL.

PREVIEW creates a deployment and creates "shell" resources but does not actually instantiate these resources. This allows you to preview what your deployment looks like. You can use this intent to preview updates to deployments or preview new deployments. You must provide a target.config with a configuration for this intent. After previewing a deployment, you can deploy your resources by making a request with the UPDATE intent or you can CANCEL the preview altogether. Note that the deployment will still exist after you cancel the preview and you must separately delete this deployment if you want to remove it.

UPDATE performs an update to the underlying resources in a deployment. If you provide a populated target.config field with this request, Deployment Manager uses that configuration to perform an update. If you had previewed this update beforehand, and do not supply a target.config or provide an empty target.config, Deployment Manager uses the last previewed configuration.

CANCEL cancels an update that is in PREVIEW or UPDATE but does not undo any changes already made.

insert_time

[Output Only] Timestamp when the deployment was created, in RFC3339 text format .

fingerprint

Specifies a fingerprint for update() requests. A fingerprint is a randomly generated value that must be provided in update() requests to perform optimistic locking. This ensures optimistic concurrency so that only one update can be performed at a time. The fingerprint is initially generated by Deployment Manager and changes after every request to modify data. To get the latest fingerprint value, perform a get() request to a deployment.

id

[Output Only] Unique identifier for the resource; defined by the server.

target

[Input Only] The parameters that define your deployment, including the deployment configuration and relevant templates.